[Bugfix] Promote sparse-MLA KV row offsets to int64
The sparse-MLA Triton kernels compute `indices * stride_kv_token` in 32-bit.
A row id fits in int32, but the element offset does not once the flat KV view
passes 2**31 / stride rows: with the NoPE 512 layout that is 4,194,304 rows,
and with the 576 layout 3,728,271. Triton evaluates the whole offset
expression before the pointer addition widens anything, so the 64-bit base
pointer does not save it -- the multiply wraps first, the load lands on a
wrong cache row, and attention returns a plausible wrong answer with no
crash and no warning.

Six sites, all of the same shape:

  * _mla_sparse_kernel: the shared K/V gather, the separate-K gather, the
    RoPE tail gather and the separate-V gather;
  * the tuned Ampere prefill kernel's clamped row term, which feeds K, K_pe
    and V from one expression;
  * the gathered prefill MQA-logits kernel's `cols * D`, whose ceiling is
    16,777,216 gathered rows.

The decode paged MQA-logits kernel and the kpool cache writers already widen
their page indices before multiplying and are unchanged.

No behaviour change below the threshold: the widened product is equal to the
narrow one wherever the narrow one was correct.

+++ vllm/v1/attention/ops/triton_mla_sparse.py
@@ -132,7 +132,7 @@ def _mla_sparse_kernel(
             # each row once in its natural [BLOCK_N, D] layout and reuse it for
             # both dots instead of reading the row twice from L2.
             offs_kn = (
-                indices[:, None] * stride_k_token
+                indices[:, None].to(tl.int64) * stride_k_token
                 + cur_kv_head_id * stride_k_head
                 + offs_d[None, :]
             )
@@ -142,7 +142,7 @@ def _mla_sparse_kernel(
             qk = tl.dot(q, tl.trans(kn))
         else:
             offs_k = (
-                indices[None, :] * stride_k_token
+                indices[None, :].to(tl.int64) * stride_k_token
                 + cur_kv_head_id * stride_k_head
                 + offs_d[:, None]
             )
@@ -155,7 +155,7 @@ def _mla_sparse_kernel(
         if BLOCK_DPE > 0:
             # q_rope @ k_rope
             offs_kpe = (
-                indices[None, :] * stride_k_token
+                indices[None, :].to(tl.int64) * stride_k_token
                 + cur_kv_head_id * stride_k_head
                 + offs_dpe[:, None]
             )
@@ -175,7 +175,7 @@ def _mla_sparse_kernel(
             # load v
             mask_v_d = offs_dv < dim_v
             offs_v = (
-                indices[:, None] * stride_v_token
+                indices[:, None].to(tl.int64) * stride_v_token
                 + cur_kv_head_id * stride_v_head
                 + offs_dv[None, :]
             )
